Removed GQLRegex, all functionality moved into GQLFA
